Moved cardex page out of src
