+++ src/codex/cards.module.scss
@@ -1,19 +1,24 @@
-@use './styles/index' as *;
+@use '../styles/index' as *;
 
 .container {
   display: flex;
   flex-direction: column;
   padding: 0.375rem 0.625rem 0.625rem;
   background-color: $color-base-bg;
+  color: $color-primary-base;
   min-height: 100vh;
+  font-size: font-size('sm');
   @include starsTexture('upper');
   @include starsTexture('lower');
 }
 
 .content {
-  background-color: $color-base-bg;
-  color: $color-primary-base;
-  font-family: Arial, sans-serif;
-  margin: 0;
+  display: flex;
+  justify-content: center;
+  gap: 2rem;
   padding: 0.5rem 1rem;
+
+  @media (max-width: $breakpoint-mobile) {
+    flex-direction: column;
+  }
 }
